+++ PhenomRaidTools/Source/Utils.lua
@@ -139,7 +139,7 @@ PRT.ConfirmationDialog = function(text, successFn, ...)
                 PRT.Core.UnregisterFrame(text)
             end)
 
-        local textLabel = PRT.Label(text, 14)
+        local textLabel = PRT.Label(text, 16)
 
         confirmationFrame:SetWidth(max(430, textLabel.label:GetStringWidth() + 50))